% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_exercise_csv}
\alias{create_exercise_csv}
\title{Create an exercise csv file}
\usage{
create_exercise_csv(file, sep = ",")
}
\arguments{
\item{file}{A string, name of a text file.}

\item{sep}{Column separator character.}
}
\value{
A string.
}
\description{
Creates an empty exercise csv file.
}
\examples{

file <- create_exercise_csv(file = tempfile(fileext = '.csv'))

}
\seealso{
Other support functions: 
\code{\link{create_exercise_data_frame}()},
\code{\link{create_exercise_excel}()},
\code{\link{read_exercise_csv}()},
\code{\link{read_exercise_excel}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
