% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_exercise_data_frame}
\alias{create_exercise_data_frame}
\title{Create an exercise data frame}
\usage{
create_exercise_data_frame()
}
\value{
A data frame.
}
\description{
Creates an empty exercise data frame.
}
\examples{

df <- create_exercise_data_frame()

}
\seealso{
Other support functions: 
\code{\link{create_exercise_csv}()},
\code{\link{create_exercise_excel}()},
\code{\link{read_exercise_csv}()},
\code{\link{read_exercise_excel}()},
\code{\link{set_pending_answers}()},
\code{\link{vector_to_string}()},
\code{\link{write_exercise_csv}()}
}
\concept{support functions}
