% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_exercise_excel}
\alias{create_exercise_excel}
\title{Create an exercise Excel file}
\usage{
create_exercise_excel(file)
}
\arguments{
\item{file}{A string, name of a text file.}
}
\value{
A string.
}
\description{
Creates an empty exercise Excel file.
}
\examples{

file <- create_exercise_excel(file = tempfile(fileext = '.xlsx'))

}
\seealso{
Other support functions: 
\code{\link{create_exercise_csv}()},
\code{\link{create_exercise_data_frame}()},
\code{\link{read_exercise_csv}()},
\code{\link{read_exercise_excel}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
