% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_questions_from.R
\name{define_questions_from_csv}
\alias{define_questions_from_csv}
\alias{define_questions_from_csv.exam}
\title{Define questions from a csv file}
\usage{
define_questions_from_csv(ex, file, sep)

\method{define_questions_from_csv}{exam}(ex, file, sep = ",")
}
\arguments{
\item{ex}{An `exam` object.}

\item{file}{A string, name of a text file.}

\item{sep}{Column separator character ("," or ";").}
}
\value{
An `exam`.
}
\description{
Each row in the text file is interpreted as a question. We only need to define
the columns that we are going to use; the rest of the columns are taken by default.
}
\details{
Both the answer and the options are formed by a vector of strings, from which
one is chosen to formulate the question and select the answer. To represent a
vector of strings in a cell, the function `vector_to_string()` is used, which
includes a separator ("<|>") between the vector elements to generate a string.
}
\examples{

rmd <- system.file("extdata/template01.Rmd", package = "rexer")
questions <- system.file("extdata/questions.csv", package = "rexer")
ex <- exam(
  rmd = rmd,
  examinees = NULL,
  instances_num = 10,
  random = TRUE,
  reorder_questions = TRUE,
  select_n_questions = NULL
) |>
  define_questions_from_csv(questions)

}
\seealso{
\code{\link{exam}}, \code{\link{vector_to_string}}

Other question definition: 
\code{\link{define_a_question}()},
\code{\link{define_questions}()},
\code{\link{define_questions_from_excel}()}
}
\concept{question definition}
