% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_questions_from.R
\name{define_questions_from_excel}
\alias{define_questions_from_excel}
\alias{define_questions_from_excel.exam}
\title{Define questions from a Excel file}
\usage{
define_questions_from_excel(ex, file, sheet_index, sheet_name)

\method{define_questions_from_excel}{exam}(ex, file, sheet_index = NULL, sheet_name = NULL)
}
\arguments{
\item{ex}{An `exam` object.}

\item{file}{A string, name of an Excel file.}

\item{sheet_index}{A number, sheet index in the workbook.}

\item{sheet_name}{A string, sheet name.}
}
\value{
An `exam`.
}
\description{
Each row in the Excel file is interpreted as a question. We only need to define
the columns that we are going to use; the rest of the columns are taken by default.
}
\details{
In addition to the file, we can indicate the sheet by its name or index. If we
do not indicate anything, it considers the first sheet.

Both the answer and the options are formed by a vector of strings, from which
one is chosen to formulate the question and select the answer. To represent a
vector of strings in a cell, the function `vector_to_string()` is used, which
includes a separator ("<|>") between the vector elements to generate a string.
}
\examples{

rmd <- system.file("extdata/template01.Rmd", package = "rexer")
questions <- system.file("extdata/questions.xlsx", package = "rexer")
ex <- exam(
  rmd = rmd,
  examinees = NULL,
  instances_num = 10,
  random = TRUE,
  reorder_questions = TRUE,
  select_n_questions = NULL
) |>
  define_questions_from_excel(questions)

}
\seealso{
\code{\link{exam}}, \code{\link{vector_to_string}}

Other question definition: 
\code{\link{define_a_question}()},
\code{\link{define_questions}()},
\code{\link{define_questions_from_csv}()}
}
\concept{question definition}
