% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{generate_document}
\alias{generate_document}
\alias{generate_document.exam}
\title{generate the exam document}
\usage{
generate_document(ex, output_format, encoding)

\method{generate_document}{exam}(ex, output_format = "pdf_document", encoding = "UTF-8")
}
\arguments{
\item{ex}{A `exam` object.}

\item{output_format}{A vector of strings.}

\item{encoding}{A string.}
}
\value{
A `exam` object.
}
\description{
generate the exam document
}
\seealso{
Other question definition: 
\code{\link{define_a_question}()},
\code{\link{define_questions}()},
\code{\link{define_questions_from_csv}()},
\code{\link{define_questions_from_excel}()}
}
\concept{question definition}
