% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{generate_pdf}
\alias{generate_pdf}
\alias{generate_pdf.exam}
\title{Generate pdf exam}
\usage{
generate_pdf(ex, encoding)

\method{generate_pdf}{exam}(ex, encoding = "UTF-8")
}
\arguments{
\item{ex}{A `exam` object.}

\item{encoding}{A string.}
}
\value{
A `exam`.
}
\description{
Generate pdf exam
}
\concept{question definition}
