% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{get_question_set}
\alias{get_question_set}
\title{Get question set}
\usage{
get_question_set(n = NULL)
}
\arguments{
\item{n}{An integer, number of questions to include, with NULL all are included.}
}
\value{
A data frame
}
\description{
Get question set
}
\examples{

q <- get_question_set()

}
