% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{interpret_a_question}
\alias{interpret_a_question}
\title{interpret a question.}
\usage{
interpret_a_question(question, exam_number, random, delivery)
}
\arguments{
\item{question}{A data frame, question.}

\item{exam_number}{An integer, exam sequence number}

\item{random}{A boolean, is random generation.}

\item{delivery}{A boolean, is delivery version.}
}
\value{
A string.
}
\description{
interpret a question.
}
\keyword{internal}
