% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{interpret_all_exercises}
\alias{interpret_all_exercises}
\title{interpret all exercise}
\usage{
interpret_all_exercises(exercises, exam_number, random, reorder, delivery)
}
\arguments{
\item{exercises}{A data frame, exercises.}

\item{exam_number}{An integer, exam sequence number}

\item{random}{A boolean, is random generation.}

\item{reorder}{A boolean, reorder exercises.}

\item{delivery}{A boolean, is delivery version.}
}
\value{
A string.
}
\description{
interpret all exercise
}
\keyword{internal}
