% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{interpret_an_exercise}
\alias{interpret_an_exercise}
\title{interpret a exercise.}
\usage{
interpret_an_exercise(exercise, exam_number, random, delivery)
}
\arguments{
\item{exercise}{A data frame, exercise.}

\item{exam_number}{An integer, exam sequence number}

\item{random}{A boolean, is random generation.}

\item{delivery}{A boolean, is delivery version.}
}
\value{
A string.
}
\description{
interpret a exercise.
}
\keyword{internal}
