% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{interpret_exercises}
\alias{interpret_exercises}
\title{interpret exercises}
\usage{
interpret_exercises(exercises, exam_number, random, reorder, delivery)
}
\arguments{
\item{exercises}{A data frame, exercises.}

\item{exam_number}{An integer, exam sequence number}

\item{random}{A boolean, is random generation.}

\item{reorder}{A boolean, reorder exercises.}

\item{delivery}{A boolean, is delivery version.}
}
\value{
A string vector.
}
\description{
interpret exercises
}
\keyword{internal}
