% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{interpret_questions}
\alias{interpret_questions}
\title{interpret questions}
\usage{
interpret_questions(questions, exam_number, random, reorder, delivery)
}
\arguments{
\item{questions}{A data frame, questions.}

\item{exam_number}{An integer, exam sequence number}

\item{random}{A boolean, is random generation.}

\item{reorder}{A boolean, reorder questions.}

\item{delivery}{A boolean, is delivery version.}
}
\value{
A string vector.
}
\description{
interpret questions
}
\keyword{internal}
