% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{num_vector}
\alias{num_vector}
\title{generates a vector of numbers in string format, for the  interval.}
\usage{
num_vector(start = 1, end)
}
\arguments{
\item{start}{An integer, start of interval.}

\item{end}{An integer, end of interval.}
}
\value{
A vector of strings.
}
\description{
generates a vector of numbers in string format, for the  interval.
}
\keyword{internal}
