% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{print_vector}
\alias{print_vector}
\title{print vector line by line.}
\usage{
print_vector(vector)
}
\arguments{
\item{vector}{An vector of strings.}
}
\value{
A string.
}
\description{
print vector line by line.
}
\keyword{internal}
