% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_csv}
\alias{read_csv}
\title{read csv in a data frame}
\usage{
read_csv(file, sep = ",")
}
\arguments{
\item{file}{A string, file name.}

\item{sep}{A character, separator.}
}
\value{
A data frame
}
\description{
read csv in a data frame
}
\seealso{
Other read functions: 
\code{\link{read_excel}()}
}
\concept{read functions}
