% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_excel}
\alias{read_excel}
\title{read Excel in a data frame}
\usage{
read_excel(file, sheet_index = NULL, sheet_name = NULL)
}
\arguments{
\item{file}{A string, file name.}

\item{sheet_index}{A number, sheet number.}

\item{sheet_name}{A string, sheet name}
}
\value{
A data frame
}
\description{
read Excel in a data frame
}
\seealso{
Other read functions: 
\code{\link{read_csv}()}
}
\concept{read functions}
