% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{read_exercise_csv}
\alias{read_exercise_csv}
\title{Read an exercise csv file}
\usage{
read_exercise_csv(file, sep = ",")
}
\arguments{
\item{file}{A string, name of a text file.}

\item{sep}{Column separator character.}
}
\value{
A data frame.
}
\description{
Reads a csv file of exercises and returns a data frame.
}
\examples{

file <- system.file("extdata/exercises.csv", package = "rexer")
df <- read_exercise_csv(file)

}
\seealso{
Other support functions: 
\code{\link{create_exercise_csv}()},
\code{\link{create_exercise_data_frame}()},
\code{\link{create_exercise_excel}()},
\code{\link{read_exercise_excel}()},
\code{\link{set_pending_answers}()},
\code{\link{vector_to_string}()},
\code{\link{write_exercise_csv}()}
}
\concept{support functions}
