% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{reduce_vector}
\alias{reduce_vector}
\title{reduce vector by sep.}
\usage{
reduce_vector(vector, sep = "\\n", italics = FALSE)
}
\arguments{
\item{vector}{A vector of strings.}

\item{sep}{A string, separator to use.}

\item{italics}{A boolean.}
}
\value{
A string.
}
\description{
reduce vector by sep.
}
\keyword{internal}
