% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_random}
\alias{select_random}
\title{randomly selects the number of elements. If null, returns them all randomly reordered.}
\usage{
select_random(vector, n = NULL)
}
\arguments{
\item{vector}{A vector of strings.}

\item{n}{An integer, number of elements to select.}
}
\value{
A vector of strings.
}
\description{
randomly selects the number of elements. If null, returns them all randomly reordered.
}
