% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_sequential}
\alias{select_sequential}
\title{sequentially select an element, the one that corresponds according to the number
passed as a parameter.}
\usage{
select_sequential(vector, n = 0)
}
\arguments{
\item{vector}{A vector of strings.}

\item{n}{An integer, number in sequence.}
}
\value{
A string.
}
\description{
sequentially select an element, the one that corresponds according to the number
passed as a parameter.
}
\keyword{internal}
