% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_questions_from.R
\name{define_a_question}
\alias{define_a_question}
\alias{define_a_question.exam}
\title{Define a question}
\usage{
define_a_question(ex, type, question, image, image_alt, answer, ...)

\method{define_a_question}{exam}(
  ex,
  type = "",
  question = "",
  image = "",
  image_alt = "",
  answer = "",
  ...
)
}
\arguments{
\item{ex}{A `exam` object.}

\item{type}{A string, question type (if needed).}

\item{question}{A string, statement of the question.}

\item{image}{A string, optional, image file to include in the question.}

\item{image_alt}{A string, description of the image to include in the question.}

\item{answer}{A string, correct answer to the question.}

\item{...}{A string, rest of the answers to the question.}
}
\value{
A `exam`.
}
\description{
Define a question and the possible answers.
}
\details{
If we include an image in the question, we must also include text in the `alt`
field associated with it.

After the correct answer, we can indicate as many answers as we want, if we do
not indicate all the parameters, we have to give each answer a parameter name
different from the rest of the parameter names.
}
\seealso{
Other question definition: 
\code{\link{define_questions}()},
\code{\link{define_questions_from_csv}()},
\code{\link{define_questions_from_excel}()},
\code{\link{generate_pdf}()}
}
\concept{question definition}
