% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_questions_from.R
\name{define_a_question}
\alias{define_a_question}
\alias{define_a_question.exam}
\title{Define a question}
\usage{
define_a_question(ex, type, question, image, image_alt, answer, ...)

\method{define_a_question}{exam}(
  ex,
  type = "",
  question = "",
  image = "",
  image_alt = "",
  answer = "",
  ...
)
}
\arguments{
\item{ex}{An `exam` object.}

\item{type}{A string, question type (if needed).}

\item{question}{A string, statement of the question.}

\item{image}{A string, optional, image file to include in the question.}

\item{image_alt}{A string, description of the image to include in the question.}

\item{answer}{A string, correct answer to the question.}

\item{...}{A string, options for the gaps in the question.}
}
\value{
An `exam`.
}
\description{
Define a question with random components.
}
\details{
If we include an image in the question, we must also include text in the `alt`
field associated with it.

After the answer, the options can be defined to fill in the gaps that have been
defined in the question.
}
\seealso{
Other question definition: 
\code{\link{define_questions}()},
\code{\link{define_questions_from_csv}()},
\code{\link{define_questions_from_excel}()}
}
\concept{question definition}
