% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_questions_from.R
\name{define_questions}
\alias{define_questions}
\alias{define_questions.exam}
\title{Define questions from a data frame}
\usage{
define_questions(ex, df)

\method{define_questions}{exam}(ex, df)
}
\arguments{
\item{ex}{A `exam` object.}

\item{df}{A data frame.}
}
\value{
A `exam`.
}
\description{
Each row in the text data frame is interpreted as a question. We only have to define
the columns that we are going to use, the rest of the columns are taken by default.
}
\details{
For answers where a vector is required, "<|>" is used as a separator of the vector
elements.
}
\seealso{
Other question definition: 
\code{\link{define_a_question}()},
\code{\link{define_questions_from_csv}()},
\code{\link{define_questions_from_excel}()},
\code{\link{generate_pdf}()}
}
\concept{question definition}
