% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam}
\alias{exam}
\title{`exam` S3 class}
\usage{
exam(
  rmd = NULL,
  examined = NULL,
  instances_num = 1,
  random = TRUE,
  reorder_questions = TRUE,
  select_n_questions = NULL
)
}
\arguments{
\item{rmd}{A string, rmd file, exam template.}

\item{examined}{A vector, instance names to generate.}

\item{instances_num}{An integer, number of instances to generate, if the names
of examined are not indicated.}

\item{random}{A boolean, random or sequential generation.}

\item{reorder_questions}{A boolean, reorder questions in exam.}

\item{select_n_questions}{An integer, number of questions to include.}
}
\value{
A `exam` object.
}
\description{
Creates a `exam` object.
}
\concept{exam definition}
