% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam}
\alias{exam}
\title{`exam` S3 class}
\usage{
exam(
  rmd = NULL,
  examinees = NULL,
  instances_num = 1,
  random = TRUE,
  reorder_questions = TRUE,
  select_n_questions = NULL
)
}
\arguments{
\item{rmd}{A string, rmd file, exam template.}

\item{examinees}{A vector, instance names to generate.}

\item{instances_num}{An integer, number of instances to generate, if the names
of examinees are not indicated.}

\item{random}{A boolean, random or sequential generation.}

\item{reorder_questions}{A boolean, reorder questions in exam.}

\item{select_n_questions}{An integer, number of questions to include.}
}
\value{
A `exam` object.
}
\description{
Creates a `exam` object.
}
\details{
A test is based on an Rmd template that has predefined parameters, and their
values are filled in using the functions of this object. The `rmd` parameter
specifies the template file.

From the template, we generate multiple instances of the test. We can specify
the instances to generate in two ways: by indicating a vector of examinee names
(via the `examinees` parameter) or by specifying the number of instances to
generate (via the `instances_num` parameter). If both are indicated, the examinee
names take precedence.
}
\concept{exam definition}
