% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam}
\alias{exam}
\title{`exam` S3 class}
\usage{
exam(
  rmd = NULL,
  questions = NULL,
  students = NULL,
  instances_num = 1,
  random = TRUE,
  reorder_questions = TRUE,
  delivery = TRUE,
  out_dir = "exam",
  seed = 173
)
}
\arguments{
\item{rmd}{A string, rmd file, exam template.}

\item{questions}{A data frame, exam questions.}

\item{students}{A vector, instance names to generate.}

\item{instances_num}{An integer, number of instances to generate, if the names
of students are not indicated.}

\item{random}{A boolean, random or sequential generation.}

\item{reorder_questions}{A boolean, reorder questions in exam.}

\item{delivery}{A boolean, version to correct or delivery.}

\item{out_dir}{A string, output folder.}

\item{seed}{An integer, seed to generate random numbers.}
}
\value{
A `exam` object.
}
\description{
Creates a `exam` object.
}
\concept{exam definition}
