% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{generate_document}
\alias{generate_document}
\alias{generate_document.exam}
\title{generate the exam document}
\usage{
generate_document(ex, out_dir, output_format, encoding, pages)

\method{generate_document}{exam}(
  ex,
  out_dir = NULL,
  output_format = "pdf_document",
  encoding = "UTF-8",
  pages = NULL
)
}
\arguments{
\item{ex}{A `exam` object.}

\item{out_dir}{A string, output folder.}

\item{output_format}{A vector of strings.}

\item{encoding}{A string.}

\item{pages}{A string, with the values all, none or NULL.}
}
\value{
A `exam` object.
}
\description{
generate the exam document
}
\seealso{
Other question definition: 
\code{\link{define_a_question}()},
\code{\link{define_questions}()},
\code{\link{define_questions_from_csv}()},
\code{\link{define_questions_from_excel}()},
\code{\link{generate_correction_document}()}
}
\concept{question definition}
